refactor(group-lifecycle): give the activation status axis its own directory
The slice took packages/shared/api/group-lifecycle to 22 direct files against
a review threshold of 20, and the resolve prefix to five. The checker asks for
a feature-ownership judgement rather than a folder, and the answer is that the
five files moved here are one feature: what a group's observed connectivity
is, and when we may say so. The stored record, the coverage reading with its
evidence watermark, the banding with hysteresis, the write-or-arm decision and
the condition axis all exist to answer that one question, and nothing left
behind depends on them except through imports.

Directory density falls to 17 and the resolve cluster to 4, so both findings
clear rather than only the one this slice tripped. Considered and rejected:
merging the two banding files, which is cohesive but reaches 21 and would be a
change made to move a counter.
