+++ packages/shared/api/group-lifecycle/activation-status/group-activation-coverage-hysteresis.ts
