+++ packages/shared/api/group-lifecycle/group-formation-view.ts
@@ -3,8 +3,8 @@ import type { GroupRef } from '../group-types.ts';
 import type {
     GroupActivationCondition,
     GroupActivationRemediation
-} from './compute-group-activation-condition.ts';
-import type { GroupFormationReadiness } from './compute-group-formation-reading.ts';
+} from './activation-status/compute-group-activation-condition.ts';
+import type { GroupFormationReadiness } from './activation-status/compute-group-formation-reading.ts';
 import type { GroupLayoutIdentity } from './group-layout-identity.ts';
 import type { GroupFormationOutcome, GroupLifecycleState } from './group-lifecycle-policy.ts';
